move out dummy migration completely from src
