+++ phpunit.xml.dist
@@ -91,6 +91,7 @@
             <directory suffix=".php">src/Core/DevOps/Test</directory>
             <directory suffix=".php">src/Core/Framework/Test</directory>
             <directory suffix=".php">src/Core/Migration/Test</directory>
+            <directory suffix=".php">src/Core/Migration/Traits</directory>
             <directory suffix=".php">src/Core/Test</directory>
             <directory suffix=".php">src/Elasticsearch/Test</directory>
             <directory suffix=".php">src/Storefront/Test</directory>
